# api/api.yaml
openapi: 3.0.3
info:
  description: 'Spec for the CanisMajor DLT adaptor'
  version: 0.0.1
  title: CanisMajor - FIWARE DLT Adaptor
  contact:
    email: stefan.wiedemann@fiware.org
...
paths:
...
  '/ngsi-ld/v1/entities/':
    post:
      tags:
        - NGSI-LD
      parameters:
        - $ref: '#/components/parameters/dltToken'
        - $ref: '#/components/parameters/link'
      description: Entity creation requests.
      operationId: createNgsiLDEntity
      requestBody:
        content:
          application/ld+json:
            schema:
              $ref: '#/components/schemas/Entity'
          application/json:
            schema:
              $ref: '#/components/schemas/Entity'
      responses:
        '201':
          description: NGSI-LD Entity created.
        '400':
          description: Did not receive a valid NGSI-LD Entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: No valid token was provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Entity already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/ngsi-ld/v1/entities/{entityId}/attrs':
    post:
      tags:
        - NGSI-LD
      parameters:
        - $ref: '#/components/parameters/link'
        - $ref: '#/components/parameters/dltToken'
        - $ref: '#/components/parameters/entityId'
      description: EntityUpdate requests.
      operationId: postUpdateNgsiLDEntity
      requestBody:
        content:
          application/ld+json:
            schema:
              $ref: '#/components/schemas/EntityFragment'
          application/json:
            schema:
              $ref: '#/components/schemas/EntityFragment'
      responses:
        '204':
          description: All attributes where apended successfully.
        '207':
          description: Only attributes in the response body where appended.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: Did not receive a valid NGSI-LD Entity-Fragement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: No such entity exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  parameters:
...
    link:
      name: Link
      description: Link header to be used as described in the json-ld spec.
      in: header
      required: false
      schema:
        type: string
    dltToken:
      name: dlt-token
      description: Header to contain the dlt-token
      in: header
      required: true
      schema: 
        type: string
    entityId:
      name: entityId
      description: ID of an NGSI-LD entity
      in: path
      required: true
      schema:
        type: string
        format: uri
  schemas:
...
    positionDefinition:
      description: A single position
      type: array
      minItems: 2
      maxItems: 2
      items:
        type: number
      additionalProperties: false
    positionArrayDefinition:
      description: An array of positions
      type: array
      items:
        $ref: '#/components/schemas/positionDefinition'
    lineStringDefinition:
      description: An array of two or more positions
      type: array
      items:
        $ref: '#/components/schemas/positionDefinition'
      minItems: 2
    linearRingDefinition:
      description: An array of four positions where the first equals the last
      type: array
      items:
        $ref: '#/components/schemas/positionDefinition'
      minItems: 4
    polygonDefinition:
      description: An array of linear rings
      type: array
      items:
        $ref: '#/components/schemas/linearRingDefinition'
...
    Point:
      type: object
      properties:
        type:
          type: string
          enum:
            - Point
        coordinates:
          $ref: '#/components/schemas/positionDefinition'
    MultiPoint:
      type: object
      properties:
        type:
          type: string
          enum:
            - MultiPoint
        coordinates:
          $ref: '#/components/schemas/positionArrayDefinition'
    Polygon:
      type: object
      properties:
        type:
          type: string
          enum:
            - Polygon
        coordinates:
          $ref: '#/components/schemas/polygonDefinition'
    LineString:
      type: object
      properties:
        type:
          type: string
          enum:
            - LineString
        coordinates:
          $ref: '#/components/schemas/lineStringDefinition'
    MultiLineString:
      type: object
      properties:
        type:
          type: string
          enum:
            - MultiLineString
        coordinates:
          type: array
          items:
            $ref: '#/components/schemas/lineStringDefinition'
    MultiPolygon:
      type: object
      properties:
        type:
          type: string
          enum:
            - MultiPolygon
        coordinates:
          type: array
          items:
            $ref: '#/components/schemas/polygonDefinition'
    Geometry:
      description: ' Avalid GeoJSON geometry object'
      oneOf:
        - $ref: '#/components/schemas/Point'
        - $ref: '#/components/schemas/MultiPoint'
        - $ref: '#/components/schemas/Polygon'
        - $ref: '#/components/schemas/LineString'
        - $ref: '#/components/schemas/MultiLineString'
        - $ref: '#/components/schemas/MultiPolygon'
    observedAt:
      type: string
      format: date-time
    createdAt:
      type: string
      format: date-time
    modifiedAt:
      type: string
      format: date-time
    datasetId:
      type: string
      format: uri
    instanceId:
      type: string
      format: uri
    ProblemDetails:
      type: object
      description: Information about the occured error.
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
      required:
        - type
    GeoProperty:
      type: object
      properties:
        type:
          type: string
          enum:
            - GeoProperty
        value:
          type: object
        observedAt:
          $ref: '#/components/schemas/observedAt'
        createdAt:
          $ref: '#/components/schemas/createdAt'
        modifiedAt:
          $ref: '#/components/schemas/modifiedAt'
        datasetId:
          $ref: '#/components/schemas/datasetId'
        instanceId:
          $ref: '#/components/schemas/instanceId'
        unitCode:
          type: string
      required:
        - type
        - value
      additionalProperties:
        oneOf:
          - $ref: '#/components/schemas/Geometry'
          - $ref: '#/components/schemas/Property'
          - $ref: '#/components/schemas/Relationship'
    LdContext:
      type: object
    Property:
      type: object
      properties:
        type:
          type: string
          enum:
            - Property
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
            - type: array
              items:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
                  - type: object
        observedAt:
          $ref: '#/components/schemas/observedAt'
        createdAt:
          $ref: '#/components/schemas/createdAt'
        modifiedAt:
          $ref: '#/components/schemas/modifiedAt'
        datasetId:
          $ref: '#/components/schemas/datasetId'
        instanceId:
          $ref: '#/components/schemas/instanceId'
        unitCode:
          type: string
      required:
        - type
        - value
      additionalProperties:
        oneOf:
          - $ref: '#/components/schemas/Property'
          - $ref: '#/components/schemas/Relationship'
    Relationship:
      type: object
      properties:
        type:
          type: string
          enum:
            - Relationship
        object:
          type: string
          format: uri
        observedAt:
          $ref: '#/components/schemas/observedAt'
        createdAt:
          $ref: '#/components/schemas/createdAt'
        modifiedAt:
          $ref: '#/components/schemas/modifiedAt'
        datasetId:
          $ref: '#/components/schemas/datasetId'
        instanceId:
          $ref: '#/components/schemas/instanceId'
      required:
        - type
        - object
      additionalProperties:
        oneOf:
          - $ref: '#/components/schemas/Property'
          - $ref: '#/components/schemas/Relationship'
    Name:
      type: string
      pattern: '^((\d|[a-zA-Z]|_)+(:(\d|[a-zA-Z]|_)+)?(#\d+)?)$'
      minLength: 1
      description: NGSI-LD Name
    EntityFragment:
      type: object
      properties:
        '@context':
          $ref: '#/components/schemas/LdContext'
        location:
          $ref: '#/components/schemas/GeoProperty'
        observationSpace:
          $ref: '#/components/schemas/GeoProperty'
        operationSpace:
          $ref: '#/components/schemas/GeoProperty'
        id:
          type: string
          format: uri
        type:
          $ref: '#/components/schemas/Name'
        createdAt:
          $ref: '#/components/schemas/createdAt'
        modifiedAt:
          $ref: '#/components/schemas/modifiedAt'
      additionalProperties:
        oneOf:
          - $ref: '#/components/schemas/Property'
          - $ref: '#/components/schemas/Relationship'
          - $ref: '#/components/schemas/GeoProperty'
    Entity:
      type: object
      properties:
        '@context':
          $ref: '#/components/schemas/LdContext'
        location:
          $ref: '#/components/schemas/GeoProperty'
        observationSpace:
          $ref: '#/components/schemas/GeoProperty'
        operationSpace:
          $ref: '#/components/schemas/GeoProperty'
        id:
          type: string
          format: uri
        type:
          $ref: '#/components/schemas/Name'
        createdAt:
          $ref: '#/components/schemas/createdAt'
        modifiedAt:
          $ref: '#/components/schemas/modifiedAt'
      additionalProperties:
        oneOf:
          - $ref: '#/components/schemas/Property'
          - $ref: '#/components/schemas/Relationship'
          - $ref: '#/components/schemas/GeoProperty'
...